Fixes mp4encrypt
It fixes a bug which, in some cases, updates the tfra entries more than once. This causes "415 - Unsupported Media Type" responses from services like Akamai when the client requests  fragments with a correct moof offset but doesn't match with the tfra's entries.

So, this fix consist in prevent mp4encrypt to modify the tfra's entries of the input more than once, using a temporal mfra instead, which is a copy of the mfra at the input. Then, when the update is finished, it upgrades the mfra with the new values from the copy.