copy of file name may not be NUL-terminated
`strncpy` used a bound equal to the destination buffer size. As a consequence, an input file name of 2048 chars or more would end up being copied into buffer `fragment_input_filename` without its NUL-terminating character. This could cause stack corruption afterwards (since the buffer is subsequently searched for character '.' and returned pointer is dereferenced to write a `\0` character). This issue was detected by the compiler, through a `-Wstringop-truncation` violation warning.

Proposed fix is to simply reject filenames that don't fit, with NUL-termination, into the buffer and then simply use `strcpy()` to get rid of the warning.